+++ docs/history.rst
@@ -5,5 +5,6 @@ Release History
 
 `Next Release`_
 ---------------
+- Add :class:`sprockets.mixins.metrics.StatsdMixin`
 
 .. _Next Release: https://github.com/sprockets/sprockets.mixins.metrics/compare/0.0.0...master